added training loop

+++ sequence_dataset.py
@@ -5,7 +5,7 @@ class SequenceDataset(Dataset):
     def __init__(self, numpyArr, sequence_length=5):
         self.sequence_length = sequence_length
         # self.X = torch.from_numpy(numpyArr).float()
-        self.X = torch.from_numpy(numpyArr)
+        self.X = torch.from_numpy(numpyArr).float()
         self.size = len(numpyArr)-sequence_length
 
     def __len__(self):
